For Samples.Wcf: Refactor CustomBindingElement and other required classes into a Bindings/Custom folder, and suppress CS0067 build warning
