+++ tracer/test/test-applications/integrations/Samples.Wcf/Bindings/Custom/CustomBindingElement.cs
@@ -1,6 +1,6 @@
 using System.ServiceModel.Channels;
 
-namespace Samples.Wcf
+namespace Samples.Wcf.Bindings.Custom
 {
     public class CustomBindingElement : BindingElement
     {
