+++ tracer/test/test-applications/integrations/Samples.Wcf/Program.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.ServiceModel;
 using System.ServiceModel.Channels;
 using System.Threading.Tasks;
+using Samples.Wcf.Bindings.Custom;
 
 namespace Samples.Wcf
 {
